+++ apps/demo-manager/src/components/login-page.vue
@@ -15,7 +15,7 @@
           vc-flex-align_center
         "
       >
-        <vc-form-checkbox>Remember me</vc-form-checkbox>
+        <vc-checkbox>Remember me</vc-checkbox>
         <vc-link>Forgot your password?</vc-link>
       </div>
       <div class="vc-flex vc-flex-justify_space-between vc-flex-align_center">
@@ -57,7 +57,7 @@ export default defineComponent({
     const login = async () => {
       signInResult.value = await signIn(form.username, form.password);
       if (signInResult.value.succeeded) {
-        router.push("/orders");
+        router.replace({ name: "orders" });
       }
     };