+++ framework/ui/src/components/atoms/vc-image/vc-image.stories.ts
@@ -8,6 +8,16 @@ import VcImage from "./vc-image.vue";
 export default {
   title: "atoms/vc-image",
   component: VcImage,
+  argTypes: {
+    aspect: {
+      options: ["1x1", "3x2", "4x3", "16x9"],
+      control: { type: "radio" },
+    },
+    size: {
+      options: ["auto", "xs", "s", "m", "l", "xl"],
+      control: { type: "radio" },
+    },
+  },
 };
 
 const Template: Story = (args) => ({
@@ -22,5 +32,9 @@ const Template: Story = (args) => ({
 export const Image = Template.bind({});
 Image.storyName = "vc-image";
 Image.args = {
+  aspect: "1x1",
+  rounded: false,
+  clickable: false,
   src: "https://placekitten.com/800/600",
+  size: "auto",
 };