+++ framework/ui/src/components/atoms/vc-label/vc-label.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="vc-font-weight_bold">
+    <slot></slot>
+  </div>
+</template>