+++ framework/ui/src/components/atoms/vc-switch/vc-switch.vue
@@ -1,0 +1,3 @@
+<template>
+  <div>vc-switch</div>
+</template>